Refactor: Add centralized error handling middleware

+++ controllers/clothingItems.js
@@ -1,77 +1,59 @@
 const mongoose = require("mongoose");
+const BadRequestError = require("../errors/BadRequestError");
+const NotFoundError = require("../errors/NotFoundError");
+const ForbiddenError = require("../errors/ForbiddenError");
 const ClothingItem = require("../models/clothingItem");
-const {
-  BAD_REQUEST,
-  NOT_FOUND,
-  INTERNAL_SERVER_ERROR,
-  OK,
-  CREATED,
-  FORBIDDEN,
-} = require("../utils/errors");
+
+const { OK, CREATED } = require("../utils/errors");
 
 // Create a clothing item
-const createItem = (req, res) => {
+const createItem = (req, res, next) => {
+  const { name, weather, imageUrl } = req.body;
   const owner = req.user._id;
 
-  const { name, weather, imageUrl } = req.body;
-
-  if (!name || !weather || !imageUrl) {
-    return res.status(BAD_REQUEST).send({
-      message: "Name, weather, and imageUrl are required",
-    });
-  }
-
-  return ClothingItem.create({ name, weather, imageUrl, owner })
+  ClothingItem.create({ name, weather, imageUrl, owner })
     .then((item) => {
       res.status(CREATED).send({ data: item });
     })
     .catch((err) => {
       console.error(err);
       if (err.name === "ValidationError") {
-        return res.status(BAD_REQUEST).send({
-          message: err.message,
-        });
+        next(new BadRequestError("Invalid data provided"));
+      } else {
+        next(err);
       }
-      return res.status(INTERNAL_SERVER_ERROR).send({
-        message: "An error has occurred on the server while creating the item",
-      });
     });
 };
 
 // GET all clothing items
 
-const getItems = (req, res) => {
+const getItems = (req, res, next) => {
   ClothingItem.find({})
-    .then((items) => {
-      res.status(OK).send({ data: items });
-    })
-    .catch((err) => {
-      console.error(err);
-      return res.status(INTERNAL_SERVER_ERROR).send({
-        message: "An error has occurred on the server while getting the items",
-      });
-    });
+    .then((items) => res.send({ data: items }))
+    .catch(next); // Let middleware handle any errors
 };
-
 
 // Delete a clothing item
 
-const deleteItem = (req, res) => {
+const deleteItem = (req, res, next) => {
   const { itemId } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(itemId)) {
+    return next(new BadRequestError("Invalid item ID format"));
+  }
+
   const userId = req.user._id;
 
   return ClothingItem.findById(itemId)
     .then((item) => {
       if (!item) {
-        res.status(NOT_FOUND).send({ message: "Item not found" });
-        return Promise.reject(new Error("Item not found"));
+        throw new NotFoundError("Item not found");
       }
 
       if (!item.owner.equals(userId)) {
-        res.status(FORBIDDEN).send({
-          message: "You don't have permission to delete this item",
-        });
-        return Promise.reject(new Error("Forbidden"));
+        throw new ForbiddenError(
+          "You don't have permission to delete this item"
+        );
       }
 
       return item.deleteOne();
@@ -84,30 +66,20 @@ const deleteItem = (req, res) => {
       return undefined;
     })
     .catch((err) => {
-      if (err.message === "Item not found" || err.message === "Forbidden") {
-        // Response already sent, do nothing
-        return undefined;
-      }
       if (err.name === "CastError") {
-        return res.status(BAD_REQUEST).send({
-          message: "Invalid item ID format",
-        });
+        return next(new BadRequestError("Invalid item ID format"));
       }
       console.error(err);
-      return res.status(INTERNAL_SERVER_ERROR).send({
-        message: "An error occurred on the server while deleting the item",
-      });
+      return next(err);
     });
 };
 
 // Adding a like to a clothing item
 
-const likeItem = (req, res) => {
+const likeItem = (req, res, next) => {
   // Check if _id is valid
   if (!mongoose.Types.ObjectId.isValid(req.params.itemId)) {
-    return res.status(BAD_REQUEST).send({
-      message: "Invalid item ID format",
-    });
+    return next(new BadRequestError("Invalid item ID format"));
   }
   return ClothingItem.findByIdAndUpdate(
     req.params.itemId,
@@ -121,23 +93,17 @@ const likeItem = (req, res) => {
     .catch((err) => {
       console.error(err);
       if (err.name === "DocumentNotFoundError") {
-        return res.status(NOT_FOUND).send({
-          message: "Requested resource not found",
-        });
+        return next(new NotFoundError("Requested resource not found"));
       }
-      return res.status(INTERNAL_SERVER_ERROR).send({
-        message: "An error has occurred on the server while liking the item",
-      });
+      return next(err);
     });
 };
 
 // Removing a like from a clothing item
-const dislikeItem = (req, res) => {
+const dislikeItem = (req, res, next) => {
   // Check if _id is valid
   if (!mongoose.Types.ObjectId.isValid(req.params.itemId)) {
-    return res.status(BAD_REQUEST).send({
-      message: "Invalid item ID format",
-    });
+    return next(new BadRequestError("Invalid item ID format"));
   }
   return ClothingItem.findByIdAndUpdate(
     req.params.itemId,
@@ -151,13 +117,9 @@ const dislikeItem = (req, res) => {
     .catch((err) => {
       console.error(err);
       if (err.name === "DocumentNotFoundError") {
-        return res.status(NOT_FOUND).send({
-          message: "Requested resource not found",
-        });
+        return next(new NotFoundError("Requested resource not found"));
       }
-      return res.status(INTERNAL_SERVER_ERROR).send({
-        message: "An error has occurred on the server while disliking the item",
-      });
+      return next(err);
     });
 };
